First version merge

+++ MailArchiver.py
@@ -6,6 +6,11 @@ from google_auth_oauthlib.flow import InstalledAppFlow
 from googleapiclient.discovery import build
 from google.auth.transport.requests import Request
 
+# CONSTANTS
+## If modifying these SCOPES, delete the file token.json.
+SCOPES = ['https://www.googleapis.com/auth/gmail.readonly']
+
+
 def read_json(json_path):
     """ Read a json file.
     Parameters:
@@ -17,14 +22,10 @@ def read_json(json_path):
         d = json.load(f)
     return d
 
-# If modifying these SCOPES, delete the file token.json.
-SCOPES = ['https://www.googleapis.com/auth/gmail.readonly']
 
 def authenticate_gmail():
+    """ Perform the Gmail authentication. """
     creds = None
-    # The file token.json stores the user's access and refresh tokens, and is
-    # created automatically when the authorization flow completes for the first
-    # time.
     if os.path.exists('token.json'):
         creds = Credentials.from_authorized_user_file('token.json')
     if not creds or not creds.valid:
@@ -38,6 +39,7 @@ def authenticate_gmail():
             token.write(creds.to_json())
     return creds
 
+
 def get_messages_with_attachments(service, user_id='me', query='has:attachment'):
     try:
         response = service.users().messages().list(userId=user_id, q=query).execute()
@@ -46,28 +48,9 @@ def get_messages_with_attachments(service, user_id='me', query='has:attachment')
     except Exception as e:
         print(f"An error occurred: {e}")
         return []
-    
-def _get_message_details(service, user_id='me', msg_id=''):
-    try:
-        message = service.users().messages().get(userId=user_id, id=msg_id).execute()
-
-        # Extract subject
-        subject = next(header['value'] for header in message['payload']['headers'] if header['name'] == 'Subject')
-
-        # Extract body (assuming it's in the 'text/plain' format)
-        body = base64.urlsafe_b64decode(message['payload']['parts'][0]['body']['data']).decode('utf-8')
-
-        # Extract attachment names
-        attachment_names = [part['filename'] for part in message['payload']['parts'] if 'filename' in part]
-
-        return subject, body, attachment_names
-
-    except Exception as e:
-        print(f"An error occurred: {e}")
-        return None, None, None
 
 
-def download_attachment(service, msg_id, user_id='me', output_dir='attachments/'):
+def download_attachment(service, msg_id, output_dir, user_id='me'):
     try:
         message = service.users().messages().get(userId=user_id, id=msg_id).execute()
         
@@ -76,29 +59,49 @@ def download_attachment(service, msg_id, user_id='me', output_dir='attachments/'
             if 'filename' in part and 'mimeType' in part:
                 if part['filename'] != '' and part['mimeType'] == 'application/pdf':
                     filename = part['filename']
-                    attachment_id = part['body']['attachmentId']
-                    attachment = service.users().messages().attachments().get(userId=user_id, messageId=msg_id, id=attachment_id).execute()
-                    file_data = base64.urlsafe_b64decode(attachment['data'].encode('UTF-8'))
-                    output_path = os.path.join(output_dir, filename)
-                    with open(output_path, 'wb') as f:
-                        f.write(file_data)
-                    print(f"Attachment '{filename}' saved to {output_path}")
+                    if not os.path.exists(os.path.join(output_dir, filename)):
+                        attachment_id = part['body']['attachmentId']
+                        attachment = service.users().messages().attachments().get(userId=user_id, messageId=msg_id, id=attachment_id).execute()
+                        file_data = base64.urlsafe_b64decode(attachment['data'].encode('UTF-8'))
+                        output_path = os.path.join(output_dir, filename)
+                        with open(output_path, 'wb') as f:
+                            f.write(file_data)
+                        print(f"Attachment '{filename}' saved to {output_path}")
     except Exception as e:
         print(f"An error occurred: {e}")
 
-def main():
-    creds = authenticate_gmail()
-    service = build('gmail', 'v1', credentials=creds)
 
+def search_and_get_attachment(service, subject, destination_folder):
+    """ Search for the desired messages and save the attachment file in the destination folder. 
+    Paramters:
+        service: Google auth service.
+        subject: String.
+        destination_folder: String. Path of the folder.
+    """
+    print("Searching for mesages: {}".format(subject))
     # Specify your search criteria (e.g., subject, sender, etc.)
-    search_query = 'subject:book-n-drive-Rechnung has:attachment'
+    search_query = 'subject:{} has:attachment'.format(subject)
     messages = get_messages_with_attachments(service, query=search_query)
+    print("\tFound {} messages...".format(len(messages)))
 
     for message in messages:
         msg_id = message['id']
-        #subject, body, attachment_names = _get_message_details(service, msg_id=msg_id)
-        #print("subject={}, attachment_names={}".format(subject, attachment_names))
-        download_attachment(service, msg_id)
+        download_attachment(service, msg_id, destination_folder)
+
 
 if __name__ == '__main__':
-    main()
+    # Load the configuration file
+    settings = read_json('config.json')
+
+    # Login in Gmail
+    creds = authenticate_gmail()
+    service = build('gmail', 'v1', credentials=creds)
+
+    # perform the seach
+    for i in list(settings.keys()):
+        subject = settings[i]['subject']
+        destination_folder = settings[i]['destination_folder']
+
+        os.makedirs(destination_folder, exist_ok=True)
+        # TODO: add time search, e.g. -w 3 in the last 3 weeks, or from the last time the script was run
+        search_and_get_attachment(service, subject, destination_folder)
